增加清空文件日志的功能 增加对未绑定多因素认证用户的引导

+++ src/pages/assets/AssetPage.tsx
@@ -467,8 +467,13 @@ const AssetPage = () => {
                                     <NButton
                                         danger={true}
                                         onClick={async () => {
-                                            await api.deleteById(selectedRowKeys.join(','));
-                                            actionRef.current?.reload();
+                                            modal.confirm({
+                                                title: t('general.delete_confirm'),
+                                                onOk: async () => {
+                                                    await api.deleteById(selectedRowKeys.join(','));
+                                                    actionRef.current?.reload();
+                                                }
+                                            })
                                         }}
                                     >{t('actions.delete')}
                                     </NButton>
